updated named vectors folder structure
