+++ test/acceptance_with_go_client/named_vectors/named_vectors_batch.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package named_vectors_tests
+package named_vectors
 
 import (
 	"context"
