+++ test/acceptance_with_go_client/named_vectors/named_vectors_objects_mixed.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package named_vectors_tests
+package named_vectors
 
 import (
 	"context"
@@ -22,7 +22,7 @@ import (
 	"github.com/weaviate/weaviate/entities/schema"
 )
 
-func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
+func testCreateSchemaWithMixedVectorizers(host string) func(t *testing.T) {
 	return func(t *testing.T) {
 		ctx := context.Background()
 		client, err := wvt.NewClient(wvt.Config{Scheme: "http", Host: host})
@@ -39,7 +39,7 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 			className := "BringYourOwnVector"
 			none1 := "none1"
 			none2 := "none2"
-			bringYourTargetVectors := []string{none1, none2}
+			mixedTargetVectors := []string{none1, none2, c11y, transformers_bq}
 			vector1a := []float32{0.1, 0.2, 0.3}
 			vector2a := []float32{-0.1001, 0.2002, -0.3003, -0.4, -0.5}
 			vector1b := []float32{0.1111, 0.4, 0.3}
@@ -64,6 +64,23 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 						},
 						VectorIndexType: "flat",
 					},
+					c11y: {
+						Vectorizer: map[string]interface{}{
+							text2vecContextionary: map[string]interface{}{
+								"vectorizeClassName": false,
+							},
+						},
+						VectorIndexType: "hnsw",
+					},
+					transformers_bq: {
+						Vectorizer: map[string]interface{}{
+							text2vecTransformers: map[string]interface{}{
+								"vectorizeClassName": false,
+							},
+						},
+						VectorIndexType:   "flat",
+						VectorIndexConfig: bqFlatIndexConfig(),
+					},
 				},
 			}
 
@@ -75,8 +92,8 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 				require.NoError(t, err)
 				assert.Equal(t, class.Class, cls.Class)
 				require.NotEmpty(t, cls.VectorConfig)
-				require.Len(t, cls.VectorConfig, 2)
-				for _, targetVector := range []string{none1, none2} {
+				require.Len(t, cls.VectorConfig, len(mixedTargetVectors))
+				for _, targetVector := range mixedTargetVectors {
 					require.NotEmpty(t, cls.VectorConfig[targetVector])
 					assert.NotEmpty(t, cls.VectorConfig[targetVector].VectorIndexType)
 					vectorizerConfig, ok := cls.VectorConfig[targetVector].Vectorizer.(map[string]interface{})
@@ -119,7 +136,7 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 						Do(ctx)
 					require.NoError(t, err)
 					require.NotNil(t, objWrapper)
-					assert.Len(t, objWrapper.Object.Vectors, 2)
+					assert.Len(t, objWrapper.Object.Vectors, 4)
 
 					objs, err := client.Data().ObjectsGetter().
 						WithClassName(className).
@@ -133,7 +150,7 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 					require.True(t, ok)
 					assert.Equal(t, tt.text, properties["text"])
 					assert.Nil(t, objs[0].Vector)
-					assert.Len(t, objs[0].Vectors, 2)
+					assert.Len(t, objs[0].Vectors, len(mixedTargetVectors))
 					for targetVector, vector := range tt.vectors {
 						require.NotNil(t, objs[0].Vectors[targetVector])
 						assert.Equal(t, vector, objs[0].Vectors[targetVector])
@@ -142,7 +159,7 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 			})
 
 			t.Run("update vectors", func(t *testing.T) {
-				beforeUpdateVectors := getVectors(t, client, className, id1, bringYourTargetVectors...)
+				beforeUpdateVectors := getVectors(t, client, className, id1, mixedTargetVectors...)
 
 				updatedVector1 := []float32{0.11111111111, 0.2222222222, 0.3333333333}
 				updatedVector2 := []float32{0.1, 0.2, 0.3, 0.4, 0.5}
@@ -157,7 +174,7 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 					WithVectors(updatedVectors).
 					Do(ctx)
 				require.NoError(t, err)
-				afterUpdateVectors := getVectors(t, client, className, id1, bringYourTargetVectors...)
+				afterUpdateVectors := getVectors(t, client, className, id1, mixedTargetVectors...)
 				for targetVector, vector := range updatedVectors {
 					assert.NotEqual(t, beforeUpdateVectors[targetVector], afterUpdateVectors[targetVector])
 					assert.Equal(t, vector, models.Vector(afterUpdateVectors[targetVector]))
@@ -165,7 +182,7 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 			})
 
 			t.Run("update vectors with merge", func(t *testing.T) {
-				beforeUpdateVectors := getVectors(t, client, className, id1, bringYourTargetVectors...)
+				beforeUpdateVectors := getVectors(t, client, className, id1, mixedTargetVectors...)
 
 				updatedVector1 := []float32{0.00001, 0.0002, 0.00003}
 				updatedVector2 := []float32{1.1, 1.2, 1.3, 1.4, 1.5}
@@ -184,52 +201,13 @@ func testCreateSchemaWithNoneVectorizer(host string) func(t *testing.T) {
 					WithVectors(updatedVectors).
 					Do(ctx)
 				require.NoError(t, err)
-				afterUpdateVectors := getVectors(t, client, className, id1, bringYourTargetVectors...)
-				for targetVector, vector := range updatedVectors {
+				afterUpdateVectors := getVectors(t, client, className, id1, mixedTargetVectors...)
+				for _, targetVector := range mixedTargetVectors {
 					assert.NotEqual(t, beforeUpdateVectors[targetVector], afterUpdateVectors[targetVector])
+				}
+				for targetVector, vector := range updatedVectors {
 					assert.Equal(t, vector, models.Vector(afterUpdateVectors[targetVector]))
 				}
-			})
-
-			t.Run("GraphQL nearVector check", func(t *testing.T) {
-				tests := []struct {
-					targetVector string
-					vector       []float32
-				}{
-					{
-						targetVector: none1,
-						vector:       []float32{0.00001, 0.0002, 0.00003},
-					},
-					{
-						targetVector: none2,
-						vector:       []float32{0.1, 0.2, 0.3, 0.4, 0.5},
-					},
-				}
-				for _, tt := range tests {
-					nearVector := client.GraphQL().NearVectorArgBuilder().
-						WithVector(tt.vector).
-						WithTargetVectors(tt.targetVector)
-					resultVectors := getVectorsWithNearVector(t, client, className, id1, nearVector, bringYourTargetVectors...)
-					for _, targetVector := range bringYourTargetVectors {
-						assert.NotEmpty(t, resultVectors[targetVector])
-					}
-				}
-			})
-
-			t.Run("update without vector", func(t *testing.T) {
-				beforeUpdateVectors := getVectors(t, client, className, id1, none1)
-
-				err := client.Data().Updater().
-					WithMerge().
-					WithClassName(className).
-					WithID(id1).
-					WithProperties(map[string]interface{}{
-						"text": "No vectorizer => Update should reuse old vector",
-					}).
-					Do(ctx)
-				require.NoError(t, err)
-				afterUpdateVectors := getVectors(t, client, className, id1, none1)
-				assert.Equal(t, beforeUpdateVectors[none1], afterUpdateVectors[none1])
 			})
 		})
 	}
