+++ test/acceptance_with_go_client/named_vectors/named_vectors_vector_index_types.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package named_vectors_tests
+package named_vectors
 
 import (
 	"context"
